Update README.md
Thanks for the script, just saw a typo and some consistency changes. 
I've extended your script to run in parallel as you suggested, and wrapped the main logic into a function with the scope parameterised so you can call it at different MG scopes or array of sub scopes for more targeted auditing and remediation. Let me know if interested, I'll stick the code on my personal repo to share.